third_party/blink/web_tests/platform/generic/external/wpt/html/browsers/the-window-object/accessing-other-browsing-contexts/indexed-browsing-contexts-03-expected.txt

This is a testharness.js-based test.
FAIL Indexed child browsing contexts assert_equals: The first child browsing context's container should be the object element. expected null but got object "[object Window]"
Harness: the test ran to completion.
